refactor: static test commands
